+++ stopPluto.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+docker-compose -f test-docker-compose.yml -p pluto_test stop